+++ packages/jobtypes/src/Employer.ts
@@ -1,10 +1,8 @@
 export interface Employer {
   id: string;
   name: string;
-  location: {
-    latitude: number;
-    longitude: number;
-  };
+  latitude: number;
+  longitude: number;
   contactName: string;
   contactPhone: string;
   contactEmail: string;